requirements.txt
asyncio
playwright
openai
markdownify
python-dotenv
langchain-core
fastapi
